extreact lookup tests into own file
